requirements.txt
Flask>=2.0.0
google-cloud-storage>=2.0.0
python-dotenv>=0.19.0
firebase-admin>=5.0.0
google-cloud-bigquery>=3.0.0
flask-restx==1.2.0
functions-framework>=3.0.0
feedparser>=6.0.0
python-dateutil>=2.8.0